Refactor calls from WFSCapabilitiesService -> CapabilitiesService
