+++ service-admin/src/main/java/org/oskari/admin/LayerCapabilitiesHelper.java
@@ -7,11 +7,11 @@ import fi.nls.oskari.map.view.util.ViewHelper;
 import fi.nls.oskari.service.OskariComponentManager;
 import fi.nls.oskari.service.ServiceException;
 import fi.nls.oskari.util.PropertyUtil;
-import fi.nls.oskari.wfs.WFSCapabilitiesService;
 import fi.nls.oskari.wms.WMSCapabilitiesService;
 import org.oskari.capabilities.CapabilitiesService;
 import org.oskari.capabilities.LayerCapabilities;
 import org.oskari.capabilities.ServiceConnectInfo;
+import org.oskari.capabilities.ogc.LayerCapabilitiesWFS;
 import org.oskari.maplayer.admin.LayerAdminJSONHelper;
 import org.oskari.maplayer.model.ServiceCapabilitiesResult;
 
@@ -36,18 +36,71 @@ public class LayerCapabilitiesHelper {
             case OskariLayer.TYPE_WMS:
                 return wmsCapabilities.getCapabilitiesResults(url, version, username, password, getSystemCRSs());
             case OskariLayer.TYPE_WFS:
-                return WFSCapabilitiesService.getCapabilitiesResults(url, version, username, password, getSystemCRSs());
+                return getCapabilitiesResultsWFS(url, version, username, password, getSystemCRSs());
             case OskariLayer.TYPE_WMTS:
                 return getCapabilitiesResultsWMTS(url, version, username, password, getSystemCRSs());
             default:
                 throw new ServiceException("Couldn't determine operation based on parameters");
         }
     }
+    public static ServiceCapabilitiesResult getCapabilitiesResultsWFS (final String url, final String version,
+                                                                        final String user, final String pw, final Set<String> systemCRSs)
+            throws ServiceException {
+        try {
+            ServiceConnectInfo info = new ServiceConnectInfo(url, OskariLayer.TYPE_WFS, version);
+            info.setCredentials(user, pw);
+            Map<String, LayerCapabilities> caps = CapabilitiesService.getLayersFromService(info);
+
+
+            List<OskariLayer> layers = caps.values().stream()
+                    .map(layer -> toOskariLayer((LayerCapabilitiesWFS) layer, url, user, pw, systemCRSs))
+                    .filter(l -> l != null)
+                    .collect(Collectors.toList());
+
+            ServiceCapabilitiesResult results = new ServiceCapabilitiesResult();
+            results.setTitle("N/A");
+            if (!layers.isEmpty()) {
+                results.setVersion(layers.get(0).getVersion());
+                results.setLayers(layers.stream()
+                        .map(l -> LayerAdminJSONHelper.toJSON(l))
+                        .collect(Collectors.toList()));
+            }
+
+            return results;
+        } catch (ServiceException e) {
+            throw e;
+        } catch (Exception e) {
+            throw new ServiceException ("Failed to get capabilities version: " + version + " from url: " + url, e);
+        }
+
+    }
+
+    private static OskariLayer toOskariLayer(LayerCapabilitiesWFS layer, String url, String user, String pw, Set<String> systemCRSs) {
+        final OskariLayer ml = new OskariLayer();
+        ml.setType(OskariLayer.TYPE_WFS);
+        ml.setUrl(url);
+        ml.setMaxScale(1d);
+        ml.setMinScale(1500000d);
+        ml.setName(layer.getName());
+        ml.setVersion(layer.getVersion());
+        ml.setPassword(pw);
+        ml.setUsername(user);
+        String title = layer.getTitle();
+        if (title == null || title.isEmpty()) {
+            title = layer.getName();
+        }
+        for (String lang : PropertyUtil.getSupportedLanguages()) {
+            ml.setName(lang, title);
+        }
+        ml.setCapabilities(CapabilitiesService.toJSON(layer, systemCRSs));
+        return ml;
+    }
 
     public static ServiceCapabilitiesResult getCapabilitiesResultsWMTS (final String url, final String version,
                                                                  final String user, final String pw, final Set<String> systemCRSs)
             throws ServiceException {
         ServiceCapabilitiesResult results = new ServiceCapabilitiesResult();
+        results.setTitle("N/A");
         results.setVersion(version);
 
         ServiceConnectInfo info = new ServiceConnectInfo(url, OskariLayer.TYPE_WMTS, version);
